Project Restructure
Taken an object oriented approach to recreate this project.

Created Classes for the cards, deck, player, dealer and game.

Each class has functions or comments describing the functionallity of the class.

Main.cpp is now clear, it contains a single instance of a unique game pointer which will be freed from memory when out of scope.